# sidecar/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "edgar-sidecar",
    "summary": "HTTP wrapper around edgartools - the single SEC egress",
    "version": "0.1.0"
  },
  "paths": {
...
    "/filings": {
      "get": {
        "summary": "List Filings",
        "operationId": "list_filings_filings_get",
        "parameters": [
          {
            "name": "year",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 2100,
                  "minimum": 1994
                },
                {
                  "type": "null"
                }
              ],
              "title": "Year"
            }
          },
          {
            "name": "quarter",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 4,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Quarter"
            }
          },
          {
            "name": "form",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "title": "Form"
            }
          },
          {
            "name": "amendments",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Amendments"
            }
          },
          {
            "name": "date_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date From"
            }
          },
          {
            "name": "date_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date To"
            }
          },
          {
            "name": "id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Id"
            }
          },
          {
            "name": "start",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Start"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 50,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FilingsPage"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          }
        }
      }
    },
    "/filings/current": {
      "get": {
        "summary": "List Current Filings",
        "operationId": "list_current_filings_filings_current_get",
        "parameters": [
          {
            "name": "form",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Form"
            }
          },
          {
            "name": "owner",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "include",
                "exclude",
                "only"
              ],
              "type": "string",
              "default": "include",
              "title": "Owner"
            }
          },
          {
            "name": "amendments",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Amendments"
            }
          },
          {
            "name": "start",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Start"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/CurrentPageSize",
              "default": 40
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentFilingsPage"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Gateway"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CurrentFilingRef": {
        "properties": {
          "accession_number": {
            "type": "string",
            "pattern": "^\\d{10}-\\d{2}-\\d{6}$",
            "title": "Accession Number"
          },
          "form": {
            "type": "string",
            "title": "Form"
          },
          "cik": {
            "type": "string",
            "pattern": "^\\d{10}$",
            "title": "Cik"
          },
          "company": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company"
          },
          "filing_date": {
            "type": "string",
            "format": "date",
            "title": "Filing Date"
          },
          "accepted": {
            "type": "string",
            "format": "date-time",
            "title": "Accepted"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "accession_number",
          "form",
          "cik",
          "company",
          "filing_date",
          "accepted"
        ],
        "title": "CurrentFilingRef"
      },
      "CurrentFilingsPage": {
        "properties": {
          "filings": {
            "items": {
              "$ref": "#/components/schemas/CurrentFilingRef"
            },
            "type": "array",
            "title": "Filings"
          },
          "start": {
            "type": "integer",
            "title": "Start"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More"
          },
          "next_start": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Start"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "filings",
          "start",
          "page_size",
          "has_more",
          "next_start"
        ],
        "title": "CurrentFilingsPage"
      },
      "CurrentPageSize": {
        "type": "integer",
        "enum": [
          10,
          20,
          40,
          80,
          100
        ],
        "title": "CurrentPageSize",
        "description": "The getcurrent feed only serves these page sizes; anything else would be silently clamped."
      },
...
      "ErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "description": "Canonical body for every non-2xx the sidecar emits (HTTPException, mapped\nedgartools/httpx exceptions, and flattened request-validation errors alike)."
      },
...
      "FilingRef": {
        "properties": {
          "accession_number": {
            "type": "string",
            "pattern": "^\\d{10}-\\d{2}-\\d{6}$",
            "title": "Accession Number"
          },
          "form": {
            "type": "string",
            "title": "Form"
          },
          "cik": {
            "type": "string",
            "pattern": "^\\d{10}$",
            "title": "Cik"
          },
          "company": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company"
          },
          "filing_date": {
            "type": "string",
            "format": "date",
            "title": "Filing Date"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "accession_number",
          "form",
          "cik",
          "company",
          "filing_date"
        ],
        "title": "FilingRef"
      },
...
      "FilingsPage": {
        "properties": {
          "filings": {
            "items": {
              "$ref": "#/components/schemas/FilingRef"
            },
            "type": "array",
            "title": "Filings"
          },
          "total": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total"
          },
          "start": {
            "type": "integer",
            "title": "Start"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More"
          },
          "next_start": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Start"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "filings",
          "total",
          "start",
          "page_size",
          "has_more",
          "next_start"
        ],
        "title": "FilingsPage"
      },
...
    }
  }
}